refactor(vue): minor reorganization and improvements
