+++ packages/vue/src/components/Menu/Menu.ts
@@ -25,26 +25,6 @@ export enum Items {
   None = 'None',
 }
 
-export enum Keys {
-  Space = ' ',
-  Enter = 'Enter',
-  Escape = 'Escape',
-  Backspace = 'Backspace',
-
-  ArrowLeft = 'ArrowLeft',
-  ArrowUp = 'ArrowUp',
-  ArrowRight = 'ArrowRight',
-  ArrowDown = 'ArrowDown',
-
-  Home = 'Home',
-  End = 'End',
-
-  PageUp = 'PageUp',
-  PageDown = 'PageDown',
-
-  Tab = 'Tab',
-}
-
 export enum MenuState {
   Open,
   Closed,
