+++ packages/vue/src/components/Menu/MenuList.ts
@@ -1,8 +1,8 @@
 import { defineComponent, onMounted } from 'vue'
-import { render } from '../../utils'
+import { Keys, render } from '../../utils'
 import { useMenuListContext } from '../../contexts'
 import { useGenerateId } from '../../composable'
-import { Items, Keys } from './Menu'
+import { Items } from './Menu'
 
 export const MenuList = defineComponent({
   name: 'MenuList',
