+++ packages/vue/src/utils/Keyboard.ts
@@ -1,0 +1,19 @@
+export enum Keys {
+  Space = ' ',
+  Enter = 'Enter',
+  Escape = 'Escape',
+  Backspace = 'Backspace',
+
+  ArrowLeft = 'ArrowLeft',
+  ArrowUp = 'ArrowUp',
+  ArrowRight = 'ArrowRight',
+  ArrowDown = 'ArrowDown',
+
+  Home = 'Home',
+  End = 'End',
+
+  PageUp = 'PageUp',
+  PageDown = 'PageDown',
+
+  Tab = 'Tab',
+}
