+++ packages/vue/src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './Keyboard'
 export * from './render'
 export * from './unrefAllRefs'
 export * from './UnableToPassPropsThroughFragmentError'
